+++ src/TableV8/styled.tsx
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020 The Ontario Institute for Cancer Research. All rights reserved
+ * Copyright (c) 2023 The Ontario Institute for Cancer Research. All rights reserved
  *
  * This program and the accompanying materials are made available under the terms of
  * the GNU Affero General Public License v3.0. You should have received a copy of the
@@ -135,9 +135,6 @@ export const TableHeader = styled(TableHeaderComp, {
   border-bottom: 1px solid ${COLORS.BORDER};
   position: relative;
   width: ${(props) => `${props.width || 1}px`};
-  &:last-of-type {
-    overflow-x: hidden; // stop resizer from adding horizontal scrollbar
-  }
   transition: box-shadow 0.3s cubic-bezier(0.175, 0.885, 0.32, 1.275);
   ${(props) =>
     props.sorted &&
